requirements.txt
torch>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.24
langchain>=0.1.13
langchain-community>=0.0.28
langchain-core>=0.1.34
openai>=1.14.3
streamlit>=1.32.2
tiktoken>=0.5.1
numpy>=1.26.4
transformers>=4.40.1
huggingface-hub>=0.23.1
datasets
accelerate>=0.26.0

